Move `is_solr_feature_enabled` to og.base where it belongs to.
