+++ opengever/base/__init__.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
-from zope.i18nmessageid import MessageFactory
 from opengever.base.interfaces import IFavoritesSettings
+from opengever.base.interfaces import ISearchSettings
 from plone import api
 from plone.i18n.locales import languages
+from zope.i18nmessageid import MessageFactory
 import csv
 
 _ = MessageFactory('opengever.base')
@@ -16,6 +17,11 @@ VERSION = 'GEVER Version %(version)s'
 def is_favorites_feature_enabled():
     return api.portal.get_registry_record(
         'is_feature_enabled', interface=IFavoritesSettings)
+
+
+def is_solr_feature_enabled():
+    return api.portal.get_registry_record(
+        'use_solr', interface=ISearchSettings)
 
 
 class OpenGeverCSVDialect(csv.excel):
