+++ opengever/contact/utils.py
@@ -1,3 +1,4 @@
+from opengever.base.security import elevated_privileges
 from opengever.base.utils import get_hostname
 from opengever.contact.interfaces import IContactFolder
 from plone import api
@@ -12,7 +13,7 @@ def hostname_cache_key(m, *args, **kwargs):
 @ram.cache(hostname_cache_key)
 def get_contactfolder_url(elevate_privileges=False):
     if elevate_privileges:
-        from opengever.base.security import elevated_privileges
+
         with elevated_privileges():
             result = api.content.find(object_provides=IContactFolder)
     else:
